working api with login

+++ backend/controllers/goal-controller.js
@@ -1,25 +1,28 @@
 const asyncHandler = require("express-async-handler");
-const Goal = require("../model/goal-model");
+const Goal = require("../models/goal-model");
+const User = require("../models/user-model");
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 // @desc Create goal
 // @route POST /api/goals
 // @access Private
 const createGoal = asyncHandler(async (req, res) => {
-  const { text } = req.body;
-
-  if (!text) {
+  if (!req.body.text) {
     res.status(400);
     throw new Error("Please add a text field");
   }
 
-  const goal = await Goal.create({ text });
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
+
   res.status(200).json(goal);
 });
 // @desc Update goal
@@ -29,9 +32,24 @@ const updateGoal = asyncHandler(async (req, res) => {
   const { id } = req.params;
   const goal = await Goal.findById(req.params.id);
 
+  // Check for goal
   if (!goal) {
     res.status(400);
     throw new Error("Please add a valid id parameter");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(id, req.body, { new: true });
@@ -45,11 +63,23 @@ const deleteGoal = asyncHandler(async (req, res) => {
   const { id } = req.params;
   const goal = await Goal.findById(id);
 
-  console.log(goal);
-
   if (!goal) {
     res.status(400);
     throw new Error("Please add a valid id parameter");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorised");
   }
 
   await goal.deleteOne();
